JP-218 (slice 1): first-class Documents surface + headless browser model
Promotes document browsing out of the Settings modal into a full-bleed,
first-class "Documents" surface (the IA fix JP-218 was filed for). It's an
app *mode*, not a modal: the editor stays mounted underneath, and you leave
Documents by opening/creating a document or the "Editor" back button.

Reuse over rebuild — extract the browser engine, then wrap it twice:
- `useDocumentBrowserModel` — headless hook holding all the filter / sort /
  group / selection / transfer / collection / offline logic that used to be
  locked inside the 1276-line `DocumentBrowser`. Adds a nav-rail
  single-collection filter.
- `DocumentList` + `SelectionBar` — presentational card area, shared by both
  chromes.
- `DocumentBrowser` is now thin Settings-tab chrome over the model (render
  preserved; the Settings tab is unchanged this slice).
- `DocumentsHome` — the new surface: sidebar (identity, nav rail
  All/Recents/Local/Cloud/Offline, Collections, a Trash placeholder pending
  relay deletion signals, local storage foot, user + theme) and main area
  (header w/ search + sort + view + New, "Continue working" strip, the shared
  list). Brand kit visual language via existing index.css tokens.

Entry points: a "Documents" toolbar button, Ctrl/Cmd+Shift+O, and a
"Go to Documents" command-palette entry (dispatches an event App listens for,
mirroring import-diagram). Storage "Manage" and the cloud identity card route
to the existing Settings tabs for now; the storage fold-in (JP-215) and the
cloud surface (JP-213) are later slices.

Verified: typecheck + full suite (1937 tests, incl. 12 new pure-logic tests
for the comparator / permission gates / transfer-error mapping). The surface
itself needs a manual dev run (and a relay for cloud docs) to E2E.
